applied libft.h, removed unneccessary static functions

+++ ft_strlen.c
@@ -9,12 +9,11 @@
 /*   Updated: 2024/12/05 20:22:47 by smalinow         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
-
-#include <stddef.h>
+#include "libft.h"
 
 size_t	ft_strlen(const char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i])
